requirements.txt
anthropic>=0.40.0
flask>=3.0.0
youtube-transcript-api>=0.6.2
python-dotenv>=1.0.0
